refactor: extract @maple/query-engine-integrations
Moves the vendor and product-surface query builders out of the core warehouse
package: Cloudflare (6 modules), PlanetScale (2), plus setup-audit,
billing-usage and internal — ~2,900 lines. The dependency now runs one way, and
nothing under integrations may be imported by a core query.

Session replay/events deliberately stayed behind. The MCP tool layer in
`@maple/query-engine/observability` builds directly on those builders
(`observability/session-replays.ts`, `session-events.ts`), so moving them would
have made the two packages circular. That is recorded in the new package's
barrel rather than left for the next person to rediscover.

Core exposes the primitives the integrations build on as explicit subpaths —
`@maple/query-engine/ch/{tables,schema,format,query-helpers}` — instead of
widening the main barrel to serve a downstream package.

The SQL catalog had to split for the same acyclicity reason: core's catalog
cannot import these builders, so the 8 vendor fixtures move to
`query-engine-integrations/src/catalog.ts` with their own baseline snapshot and
their own tenant-scope assertion. Verified the 8 moved shapes are byte-identical
to what core emitted before the move; core's baseline diff is pure removals,
with nothing added or modified, so no remaining query changed.

Consumers: 5 files in apps/api repointed at the new package (the rest reach
these builders through the `CH` namespace re-export, which moved with them).
apps/web and apps/local-ui pick up the dependency without source changes.
