+++ apps/api/package.json
@@ -43,7 +43,8 @@
 		"@tinybirdco/sdk": "catalog:tinybird",
 		"autumn-js": "^1.2.29",
 		"drizzle-orm": "^0.45.1",
-		"effect": "catalog:effect"
+		"effect": "catalog:effect",
+		"@maple/query-engine-integrations": "workspace:*"
 	},
 	"devDependencies": {
 		"@ai-sdk/openai-compatible": "^2.0.48",
