+++ packages/query-engine-integrations/package.json
@@ -1,0 +1,28 @@
+{
+	"name": "@maple/query-engine-integrations",
+	"private": true,
+	"type": "module",
+	"exports": {
+		".": "./src/index.ts",
+		"./cloudflare": "./src/cloudflare/index.ts",
+		"./planetscale": "./src/planetscale/index.ts",
+		"./product": "./src/product/index.ts",
+		"./catalog": "./src/catalog.ts"
+	},
+	"scripts": {
+		"test": "vitest run",
+		"typecheck": "tsc --noEmit"
+	},
+	"dependencies": {
+		"@maple-dev/clickhouse-builder": "workspace:*",
+		"@maple/domain": "workspace:*",
+		"@maple/query-engine": "workspace:*",
+		"effect": "catalog:effect"
+	},
+	"devDependencies": {
+		"@effect/language-service": "catalog:effect",
+		"@types/node": "catalog:tooling",
+		"typescript": "catalog:tooling",
+		"vitest": "catalog:"
+	}
+}
