+++ packages/query-engine-integrations/src/cloudflare/cloudflare-usage.ts
@@ -10,9 +10,9 @@
 import { Schema } from "effect"
 import * as CH from "@maple-dev/clickhouse-builder/expr"
 import { from, param, type CompiledQueryRowSchema } from "@maple-dev/clickhouse-builder"
-import { CHNumber } from "../schema"
-import { MetricsSum } from "../tables"
-import { ISO_Z_FORMAT, isoBucket } from "./format"
+import { CHNumber } from "@maple/query-engine/ch/schema"
+import { MetricsSum } from "@maple/query-engine/ch/tables"
+import { ISO_Z_FORMAT, isoBucket } from "@maple/query-engine/ch/format"
 
 
 /**
