+++ packages/query-engine-integrations/src/index.ts
@@ -1,0 +1,19 @@
+// ---------------------------------------------------------------------------
+// @maple/query-engine-integrations
+//
+// Query builders for third-party integrations (Cloudflare, PlanetScale) and for
+// product surfaces that read telemetry to back one specific feature rather than
+// a general signal.
+//
+// Split out of @maple/query-engine so the core package stays the generic
+// warehouse layer: the dependency runs one way only, and nothing here may be
+// imported by a core query.
+//
+// Session replay/events deliberately stayed in the core package — the MCP tool
+// layer in `@maple/query-engine/observability` builds on those queries, so
+// moving them would make the two packages circular.
+// ---------------------------------------------------------------------------
+
+export * from "./cloudflare/index"
+export * from "./planetscale/index"
+export * from "./product/index"
