+++ packages/query-engine-integrations/src/planetscale/planetscale-map.ts
@@ -22,8 +22,8 @@
 import { Schema } from "effect"
 import * as CH from "@maple-dev/clickhouse-builder/expr"
 import { from, fromQuery, param, type CompiledQueryRowSchema } from "@maple-dev/clickhouse-builder"
-import { CHNumber } from "../schema"
-import { MetricsGauge } from "../tables"
+import { CHNumber } from "@maple/query-engine/ch/schema"
+import { MetricsGauge } from "@maple/query-engine/ch/tables"
 
 /** Active connections (gauge) — one series per edge region/branch; summed per timestamp. */
 export const CONNECTION_METRIC_NAMES = [
