+++ packages/query-engine/package.json
@@ -18,7 +18,11 @@
 		"./runtime": "./src/runtime/index.ts",
 		"./observability": "./src/observability/index.ts",
 		"./local": "./src/local.ts",
-		"./sql-catalog": "./src/sql-catalog.ts"
+		"./sql-catalog": "./src/sql-catalog.ts",
+		"./ch/tables": "./src/ch/tables.ts",
+		"./ch/schema": "./src/ch/schema.ts",
+		"./ch/format": "./src/ch/queries/format.ts",
+		"./ch/query-helpers": "./src/ch/queries/query-helpers.ts"
 	},
 	"scripts": {
 		"test": "vitest run",
