+++ packages/query-engine/src/ch/builder-fixtures.ts
@@ -390,68 +390,4 @@ export const builderFixtures: ReadonlyArray<BuilderFixture> = [
 		compile: () => CH.compile(CH.activeOrgsByLogsQuery(), { startTime: START_TIME }),
 	},
 
-	// ----- cloudflare / planetscale: one shape per module, so the extraction
-	// ----- in phase 8 can be proven byte-identical.
-	{
-		module: "cloudflare-infra",
-		name: "cloudflareZoneLatencySQL",
-		label: "default",
-		compile: () => CH.compile(CH.cloudflareZoneLatencySQL(), window),
-	},
-	{
-		module: "cloudflare-infra",
-		name: "cloudflareZoneTimeseriesSQL",
-		label: "default",
-		compile: () => CH.compile(CH.cloudflareZoneTimeseriesSQL(), { ...window, bucketSeconds: 300 }),
-	},
-	{
-		// Filters exercise the zone-slice predicates the /infra/cloudflare page sends.
-		module: "cloudflare-infra",
-		name: "cloudflareZoneTimeseriesSQL",
-		label: "filtered",
-		compile: () =>
-			CH.compile(
-				CH.cloudflareZoneTimeseriesSQL({
-					hosts: ["example.com"],
-					statusClasses: ["5xx"],
-					methods: ["GET"],
-				}),
-				{ ...window, bucketSeconds: 300 },
-			),
-	},
-	{
-		module: "cloudflare-infra-extended",
-		name: "cloudflareQueueGaugesSQL",
-		label: "default",
-		compile: () => CH.compile(CH.cloudflareQueueGaugesSQL(), window),
-	},
-	{
-		module: "cloudflare-infra-breakdowns",
-		name: "cloudflareZoneBreakdownTimeseriesSQL",
-		label: "default",
-		compile: () =>
-			CH.compile(CH.cloudflareZoneBreakdownTimeseriesSQL("path", {}, ["/api/v2/traces"]), {
-				...window,
-				serviceName: "cloudflare-zone-example-com",
-				bucketSeconds: 300,
-			}),
-	},
-	{
-		module: "cloudflare-usage",
-		name: "cloudflareUsageQuery",
-		label: "default",
-		compile: () => CH.compile(CH.cloudflareUsageQuery(), { ...window, bucketSeconds: 3600 }),
-	},
-	{
-		module: "cloudflare-map",
-		name: "cloudflareServiceLatencySQL",
-		label: "default",
-		compile: () => CH.compile(CH.cloudflareServiceLatencySQL(), window),
-	},
-	{
-		module: "planetscale-map",
-		name: "planetscaleGaugesSQL",
-		label: "default",
-		compile: () => CH.compile(CH.planetscaleGaugesSQL(), window),
-	},
 ]
